requirements.txt
Django==1.11.18
django-categories==1.6
django-cors-headers==2.4.0
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.6.4
Pillow==5.4.1
python-decouple==3.1
pytz==2018.9
six==1.12.0
unicode-slugify==0.1.3
Unidecode==1.0.23
-e git+https://github.com/abhisgithub/PyEtherpadLite.git#egg=PyEtherpadLite